requirements.txt
# Core AI Framework
crewai==0.130.0
crewai-tools>=0.60.0

# API Framework
fastapi>=0.110.0
uvicorn>=0.27.0

# LangChain & LLM
langchain-openai>=0.0.5
langchain-community>=0.0.14
openai>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.2

# Database and Data Processing
pandas>=2.0.0

